requirements.txt
pyaml
bs4
lxml
spacy
scikit-learn
scipy
gensim
